chore: improve component and route names
