+++ scriptured-prayer-components/src/views/index.ts
@@ -1,6 +1,6 @@
-export * from "./Info";
-export * from "./About";
 export * from "./Home";
+export * from "./About";
+export * from "./Dashboard";
 export * from "./Login";
 export * from "./NotFound";
 export * from "./Settings";
